Configuration: * Started reworked configuration to allow for more control.
Logging:
* Made logging look a bit nicer.
* Added stacktrace in debug builds.

Refactor and optimizations.

+++ Uchu.World/Objects/Perspective/Filters/FlagFilter.cs
@@ -25,7 +25,7 @@ namespace Uchu.World.Filters
             
             if (!gameObject.TryGetComponent<CollectibleComponent>(out var collectibleComponent)) return true;
 
-            return !Collected.Contains((float) collectibleComponent.CollectibleId + ((int) gameObject.Zone.ZoneId << 8));
+            return !Collected.Contains((float) collectibleComponent.CollectibleId + (gameObject.Zone.ZoneId << 8));
         }
     }
 }
